Fixed numHingers() and fully completed section 2
numHingers() was returning a hinger cell for every "1" that was present on the board. I fixed this so it now only returns true hinger cells. This fixed loads of issues I was having with the different search algorithms. I also made changes to make the data structues more efficient and now return a faster time complexity.

Fully implemented the min_safe() search and added testing.

Created 7 testing elements that test and compare the different searches giving us plenty of information for the report.